requirements.txt
psycopg2-binary
python-dotenv
rich
passlib
pytest
pytest-mock
pytest-cov
pytest-xdist